+++ src/main/java/mutsa/sns/service/CommentService.java
@@ -62,9 +62,16 @@ public class CommentService {
         }
     }
 
-    public void deleteComment(Integer articleId, Integer commentsId) {
+    public void deleteComment(Integer articleId, Integer commentsId, String username) {
+
         List<CommentEntity> commentListByArticleId = findCommentListByArticleId(articleId);
+
         for (CommentEntity commentEntity : commentListByArticleId) {
+            //  삭제를 시도하는 사용자와 comment 작성자가 일치한지 확인.
+            if(commentEntity.getUsername().equals(username)){
+                throw new CustomException(ErrorCode.INVALID_PERMISSION);
+            }
+            // 삭제할 comment 의 Id 와 entity 의 id 를 확인.
             if(commentEntity.getId().equals(commentsId)) {
                 commentRepository.delete(commentEntity);
                 break;
